// api/generated-client/models/ListingResource.ts
export interface ListingResource {
    /**
     * 
     * @type {number}
     * @memberof ListingResource
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof ListingResource
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof ListingResource
     */
    description: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListingResource
     */
    price: number;
    /**
     * 
     * @type {number}
     * @memberof ListingResource
     */
    floor: number | null;
    /**
     * 
     * @type {number}
     * @memberof ListingResource
     */
    surface: number | null;
    /**
     * 
     * @type {number}
     * @memberof ListingResource
     */
    minDuration: number | null;
    /**
     * 
     * @type {number}
     * @memberof ListingResource
     */
    numberRooms: number | null;
    /**
     * 
     * @type {number}
     * @memberof ListingResource
     */
    numberPersons: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof ListingResource
     */
    isReady: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ListingResource
     */
    isNegotiable: boolean;
    /**
     * 
     * @type {number}
     * @memberof ListingResource
     */
    boostLevel: number;
    /**
     * 
     * @type {string}
     * @memberof ListingResource
     */
    moderationStatus: string;
    /**
     * 
     * @type {string}
     * @memberof ListingResource
     */
    image: string;
    /**
     * BelongsTo
     * @type {RentDurationResource}
     * @memberof ListingResource
     */
    rentDuration?: RentDurationResource;
    /**
     * 
     * @type {TypeResource}
     * @memberof ListingResource
     */
    type?: TypeResource;
    /**
     * 
     * @type {LocationResource}
     * @memberof ListingResource
     */
    location?: LocationResource;
    /**
     * Many To Many
     * @type {Array<CategoryResource>}
     * @memberof ListingResource
     */
    categories?: Array<CategoryResource>;
    /**
     * 
     * @type {Array<FeatureResource>}
     * @memberof ListingResource
     */
    features?: Array<FeatureResource>;
    /**
     * 
     * @type {Array<NearPlaceResource>}
     * @memberof ListingResource
     */
    nearPlaces?: Array<NearPlaceResource>;
    /**
     * 
     * @type {Array<MemberResource>}
     * @memberof ListingResource
     */
    members?: Array<MemberResource>;
    /**
     * 
     * @type {Array<ImagesResource>}
     * @memberof ListingResource
     */
    images?: Array<ImagesResource>;
    /**
     * 
     * @type {Date}
     * @memberof ListingResource
     */
    timePost: Date | null;
}